requirements.txt
curl_cffi==0.6.2
requests
python-dotenv
supabase
pandas
